chore (PR Template): Remove markdown link for Jira cards
This fixes a pet peeve of mine. Our PR template encourages filling in the Jira
card number _and_ a URL to a Jira card. But the number is already in the URL.

This _does_ make PRs slightly uglier since you see URLs instead of nice tidy
numbers. This tradeoff is worth it for less typing, IMO.